(OFBIZ-6525) Add confirmation message before deleting entity record from web-tools application.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1721089 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/webtools/webapp/webtools/entity/ListGeneric.ftl
@@ -34,7 +34,7 @@ under the License.
                         <td class="button-col">
                             <a href='<@ofbizUrl>ViewGeneric?${record.findString}</@ofbizUrl>'>${uiLabelMap.CommonView}</a>
                         <#if hasDeletePermission == 'Y'>
-                            <a href='<@ofbizUrl>UpdateGeneric?${record.findString}&amp;UPDATE_MODE=DELETE</@ofbizUrl>'>${uiLabelMap.CommonDelete}</a>
+                            <a onclick="return confirm ('${uiLabelMap.WebtoolsConfirmDelete}')" href='<@ofbizUrl>UpdateGeneric?${record.findString}&amp;UPDATE_MODE=DELETE</@ofbizUrl>'>${uiLabelMap.CommonDelete}</a>
                         </#if>
                         </td>
                         <#list fieldList as field>
